chore: move webhook utils into testutils
