+++ test/integration/suite_test.go
@@ -146,10 +146,10 @@ func TestMain(m *testing.M) {
 		fmt.Println("INFO: starting the controller manager")
 		standardControllerArgs := []string{
 			fmt.Sprintf("--ingress-class=%s", ingressClass),
-			fmt.Sprintf("--admission-webhook-cert=%s", kongSystemServiceCert),
-			fmt.Sprintf("--admission-webhook-key=%s", kongSystemServiceKey),
+			fmt.Sprintf("--admission-webhook-cert=%s", testutils.KongSystemServiceCert),
+			fmt.Sprintf("--admission-webhook-key=%s", testutils.KongSystemServiceKey),
 			fmt.Sprintf("--watch-namespace=%s", watchNamespaces),
-			fmt.Sprintf("--admission-webhook-listen=%s:%d", admissionWebhookListenHost, admissionWebhookListenPort),
+			fmt.Sprintf("--admission-webhook-listen=%s:%d", testutils.AdmissionWebhookListenHost, testutils.AdmissionWebhookListenPort),
 			"--profiling",
 			"--dump-config",
 			"--log-level=trace",
